[SCRUM-364] Refactor: 타임유틸 이동 및 편의점 리스트에서 시간 추가
