+++ src/main/java/com/kkinikong/be/convenience/dto/response/ConveniencePostListResponse.java
@@ -1,10 +1,15 @@
 package com.kkinikong.be.convenience.dto.response;
 
 import com.kkinikong.be.convenience.domain.ConveniencePost;
+import com.kkinikong.be.global.util.TimeUtil;
 
-public record ConveniencePostListResponse(Long id, String name, boolean isAvailable) {
+public record ConveniencePostListResponse(
+    Long id, String name, boolean isAvailable, String CreatedAt) {
   public static ConveniencePostListResponse from(ConveniencePost conveniencePost) {
     return new ConveniencePostListResponse(
-        conveniencePost.getId(), conveniencePost.getName(), conveniencePost.getIsAvailable());
+        conveniencePost.getId(),
+        conveniencePost.getName(),
+        conveniencePost.getIsAvailable(),
+        TimeUtil.relativeTimeFormatter(conveniencePost.getCreatedDate()));
   }
 }
